code/requirement.txt
streamlit
sqlalchemy
scikit-learn==0.21
scikit-image
imutils
opencv-contrib-python
numpy
pillow

